[native] introduce user profile component
Summary:
This diff introduces the barebones `UserProfile` component. Right now we only have the user avatar and the username, but in subsequent diffs we will be adding to this component with the various other parts needed for the profile.

Here is what the designs look like:
{F703828}

Depends on D8914

Test Plan:
Please see screenshot below

{F703826}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8914