[web] Make chat mentions clickable
Summary:
Solution for [ENG-4561](https://linear.app/comm/issue/ENG-4561/redirect-to-given-chat-when-mention-is-clicked).

This diff introduces new hook which redirects to chat when mention is clicked and uses it in `markdown-utils.js` and finally in `markdown-chat-mention.react.js`

Depends on D8852.

Test Plan:
Before testing, please apply the entire web stack.

Test if mentions inside a spoiler are properly rendered and are clickable only when spoiler is opened:
{F698073}

Reviewers: tomek, inka, rohan

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8875