[native] introduce copy username functionality to user profiles
Summary:
With profiles, we want to give the user the ability to easily copy another user's username into their clipboard. This diff introduces this functionality.

In the figma doc, @ted described how this functionality should behave:

> The “Copy username” toggle will turn into “Username copied!” to confirm that the username has been copied successfully. The toggle will revert back to the default state after 3 seconds.

Here are the screenshots of the designs:
{F703867}

Depends on D8914

Test Plan:
Please see demo video below

{F703871}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ted, ashoat, tomek

Differential Revision: https://phab.comm.dev/D8915