[native] introduce profileLarge avatar size
Summary: In the profiles design we introduce a new avatar size (90 x 90). The closest size we had that could match was `profile` but it is 112x112. Ted and I tried using 112x112, but agreed that it was too big. I figured that the best approach here would be to make the `profile` size be 90x90, and introducd `profileLarge` that would be 112x112 (the old `profile` size).

Test Plan: `flow` and confirmed in all places where we rendered an avatar with the `profile` size that there were no regressions when I updated them

Reviewers:

Subscribers: