[web][native] Modify typeahead function and variable names to distinguish chat mention utilities
Summary:
We need to distinguish user mentions from chat mentions so it's necessary to rename functions and variables which names are too generic. Also added word "mention" to typeahead tooltip utilities -> typeahead can be used not only for mentioning, so we should also indicate that.

Depends on D8897.

Test Plan: Check if there are no typos and if every affected part of code works properly.

Reviewers: tomek, inka, rohan

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8898