[native] Add a new screen to display a QR code as part of a new native login flow
Summary:
This is the `native` counterpart of D8821. I created a new navigator because this sign in via QR code flow will inevitably have multiple 'steps' in it (like choosing a device name once the QR code is
scanned), so I decided to have it done now rather than later.

One important note: The designs are not finalized, but I've spoken with @ted and I'll do another pass on the designs in a later diff once they're ready (this is tracked in [[
https://linear.app/comm/issue/ENG-4701/follow-up-on-qr-code-page-screen-designs | ENG-4701 ]].

Addresses [[ https://linear.app/comm/issue/ENG-4515/add-a-new-screen-to-display-a-qr-code-as-part-of-a-new-native-login | ENG-4515 ]]

Depends on D8870

Test Plan: Click on the button and expect the new screen to show

Reviewers: atul, ginsu, tomek, varun

Subscribers: ashoat, ted

Differential Revision: https://phab.comm.dev/D8912