[web] Introduce `ThreadEmojiAvatarSelectionModal`
Summary:
Enables setting of emoji avatars on `web`. As of this diff all varieties of thread avatar can be set from `web`. We just need to make a few layout tweaks in `ThreadSettingsModalGeneralTab` in order to get this ready to ship.

---

Depends on D8855

Test Plan: Works as expected: {F694299}

Reviewers: ashoat, ginsu

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8856