[reports-service] Add HTTP server
Summary:
A Hello World HTTP server with middleware for CORS and logging.
Increased JSON payload size limit to 10MB because reports might be large.

Depends on D8865

Test Plan:
Used Postman to send GET request to `/health` endpoint. It returned 200 OK and empty body.
A built-in request log was printed to console.

Reviewers: michal, patryk, jon

Reviewed By: jon

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8947