[keyserver] Simplify fetching threads
Summary: All the other fetchers return a collection directly without intermediate object - we can do the same for threads so that the code becomes simpler and more simplifications could be made further in the stack.

https://linear.app/comm/issue/ENG-4654/unify-interfaces-in-session-utils

Depends on D8941

Test Plan: Modify threads in the keyserver db and check if the state is correctly fixed.

Reviewers: kamil, bartek, inka