[native] introduce barebones bottom sheet component
Summary: This components introduces the barebones bottom sheet component. I am very much working off the idea that this is far from the finished bottom sheet component, and some stuff here might not look super polished. I thought for review this would be easier to follow the progression of this component as I add the bits and pieces to it.

Test Plan: Please see the demo video to see how the bottom sheet looks/behaves so far

Reviewers:

Subscribers: