[keyserver] Determine which infos have invalid hashes based on specs
Summary:
Check if a key is top level - it means that there are infos in the collection which aren't synchronized. Iterate over them and compute hashes for each item.
When a spec doesn't specify `innerHashKey`, return the values immediately - currently we do that only for `currentUserInfo` because there's exactly one current user.

https://linear.app/comm/issue/ENG-4631/migrate-checkstate-from-session-utils-to-a-spec

Depends on D8977

Test Plan: Modify keyserver db and check if the store get fixed.

Reviewers: kamil, bartek, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8979