[reports-service] E-mail: download links and JSON preview
Summary:
It is convenient to access the report directly from an e-mail.
A report preview is added (I decided to truncate it to 100 lines to avoid sending too long e-mails)

Also added two links:
- To view the entire report JSON
- To download Redux Devtools import (available only for error reports)

Depends on D8984

Test Plan:
Created a temporary HTTP endpoint that renders this HTML and opened it in the browser.
Screenshots of the final result in D8988

Reviewers: michal, patryk, jon

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8985