[native] Convert `ConnectedMessage` class component to functional `Message`
Summary:
Based on profiling of "navigating to thread w/ varun" scenario, the re-renders of `Message` take ~100ms. There's a few nested class components here, so converting to functional to try to memoize things in a more specific way to avoid re-renders.

Test Plan:
`Message` component continues to appear as expected and we're skipping render of a few child components (not in a super impactful way, however)

Reviewers: ginsu, rohan, tomek

Subscribers: