[keyserver] Determine which infos to send based on specs
Summary:
Define a new prop in a spec where we describe which keys are used when sending delete or new infos. The only complicated state to sync is `userInfos` because it contains an additional delete condition - when a user no longer has a name.

https://linear.app/comm/issue/ENG-4631/migrate-checkstate-from-session-utils-to-a-spec

Depends on D8979

Test Plan: Modify keyserver db and check if the store get fixed (including editing, inserting and deleting infos).

Reviewers: kamil, bartek, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9008