[lib][native] Stop persisting connection
Summary:
issue: https://linear.app/comm/issue/ENG-4460/refactor-connection-field
After D9018 there is no need to persist `connection` field, since we change it back to the default value on rehydrate action anyway. Now we won't be persisting it, so we also won't have the
need to be changing it on rehydrate action.

Test Plan:
Checked that `connection` is not getting rehydrated, chceked that after @@INIT action the state of `connection` is equal to the dafault value.
Checked that the `sub state for key `connection` modified, skipping.` warning is gone

Reviewers: michal, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9020