[lib][native][web] Use connection from keyserverStore
Summary:
issue: https://linear.app/comm/issue/ENG-4460/refactor-connection-field
This diff is the same as D8931 apart from the fact that we no longer need to introduce any changes in `reduceConnectionInfo`

Test Plan: ran `yarn flow-all`. Checked that the app works, and the "Disconnected" bar is not present when app is connected, and is present when keyserver is down.

Reviewers: michal, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9022