[keyserver][lib][native][web] Remove connection field from redux
Summary:
issue: https://linear.app/comm/issue/ENG-4460/refactor-connection-field
This diff is the same as D8932, but this time we don't need a migration because `connection` is not being persisted. This diff does make it temporarily be presisted, becuse all of
keyserverStore is being persisted. I will fix this in the next diff

Test Plan:
ran `yarn flow-all`.
Reviewers:
michal, kamil
Subscribers: