[native] Stop persisting connection in keyserverStore
Summary:
issue: https://linear.app/comm/issue/ENG-4460/refactor-connection-field
We don't want to persist the `connection` field, but we do want to persist the rest of `keyserverStore`.

Test Plan:
Checked that the `in` and `out` actions work as expected, checked that in rehydrated `keyserverStore` `connection` has the value of the default connection
info.

Reviewers: michal, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9024