[terraform] Set up ECS Service Connect
Summary:
We need our services talking to each other, and ECS Service Connect is the best way to do that.
Without this, services would have to talk via the public internet, which is not ideal because the traffic would go as follows:
```
client (in ECS cluster) -> load balancer -> public internet -> load balancer -> service (in the same cluster)
```
Also, we pay for Load Balancer traffic, so this would be a waste of money.

ECS Service Connect allows us to talk to services in the same cluster without going through the public internet, by using names similiar to these in Docker Compose networks: `http://blob-service:50053`

This diff enables the ECS service connect for Blob and Identity service - these two will need to be accessible for others. Also set some variables to make them easy to use.

Test Plan: Deployed this to staging, along with reports service (a blob client) with config introduced in one of the next diffs.

Reviewers: kamil, jon, varun, tomek, bartek

Reviewed By: varun, tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9038