[reports-service] Encrypt reports
Summary:
[[ https://linear.app/comm/issue/ENG-4812/consider-encrypting-reports | ENG-4812 ]].
Added encryption/decryption step for reports. It can be disabled via CLI flag (useful for debugging/local dev).

Depends on D9034, D8988

Test Plan: Stored a report in the database, then retrieved it via the API. Verified that the report was encrypted in the database and decrypted when retrieved.

Reviewers: michal, varun, patryk, jon

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9035