[native] Fix long ID display on crash screen
Summary:

On crash screen, the ID of the error report is very long and is not displayed correctly. This patch fixes the issue by moving it to a new line.

We could possibly replace UUID with [[ https://github.com/ai/nanoid#nano-id | Nano ID ]] on server side to generate shorter IDs.

Test Plan: TBD

Reviewers: inka, tomek, kamil, patryk

Subscribers: