[native] Fix long ID display on crash screen
Summary:
On crash screen, the ID of the error report is very long and is not displayed correctly. This patch fixes the issue by moving it to a new line.

NOTE: We could possibly replace UUID with [[ https://github.com/ai/nanoid#nano-id | Nano ID ]] on server side to generate shorter IDs. They might be more user friendly

Test Plan:
| **Before** | **After** |
| {F734872} | {F734875} |

Reviewers: inka, tomek, kamil, patryk

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9045