[native] Memoize `PanGestureHandler` props in `SwipeableMessage`
Summary:
Based on flamegraph from profiler, it looks like A. creating `SwipeSnake`s and `PanGestureHandler` in `SwipeableMessage` is expensive B. because we're not memoizing things as carefully as we could be we're eating this cost multiple times.

In this diff I just memoize the `PanGestureHandler` props. In the subsequent diff I'll add some more memoization to this component specifically wrt the `SwipeSnake`s

Test Plan:
"Before":

Will do another round of profiling after the subsequent diff to show improvement.

Reviewers: ginsu, tomek, rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9053