[web] Start persisting keyserverStore
Summary:
issue: https://linear.app/comm/issue/ENG-4794/start-persisting-the-keyserverstore
We want to persist the keyserverStore on web, apart from `connection` and `updatesCurrentAsOf` fields. `updatesCurrentAsOf` doesn't have a straightforward defult value - the keyserver returns
the current time or 0 depending on whether the user has accepted the policies. That's why we are using value passed from the keyserver, stored in `preloadedState`.

Test Plan:
Checked that keyserverStore is persisted - logged the `action.payload` of `rehydrateActionType` in `redux-setpu.js`. Checked that the `connceciotn` and `updatesCurrentAsOf` are set to the
values defined in `out` function of the Transform object. Chekced that is the user has to accept the policies, the value set to `updatesCurrentAsOf` is `0`, and the current time otherwise.
Checked that the application gets populated with data when the user accpest the policies (it doen't happen if the value of `updatesCurrentAsOf` is set to the current time right away).

Reviewers: michal, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9062