[lib] Add fetch timeout to reports service calls
Summary:
Raised by @tomek in D9043 - this adds a timeout to the fetch calls to the reports service.

After doing some researc, I've decided to go with `AbortController` instead of `Promise.race`. The main advantage of this solution is, that it cancels the fetch request, instead of just ignoring it. This should prevent the request from being executed in the background and potentially consuming bandwidth.

Depends on D9043

Test Plan: TBD

Reviewers: tomek!, kamil, michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9073