[lib] introduce useUserProfileThreadInfo hook
Summary: The purpose of this hook is to take in a `UserInfo` and return a `UserProfileThreadInfo`. A `UserProfileThreadInfo` is an object that contains a `PERSONAL` and `PRIVATE` `threadInfo` and an optional `pendingPersonalThreadUserInfo`.

Test Plan: Confirmed that I was able to navigate to the correct thread and I could send friend requests (the introduction of these buttons will happen in subsequent diffs)

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9247