[native] introduce onClosed callback prop to BottomSheet component
Summary: In subsquent diffs we will introduce some logic where we want to navigate back from the modal bottom sheet screen back to the screen we oringally were before we opened (navigated) to the modal bottom sheet screen when the bottomsheet closes. Unfortantly, the bottom sheet libary does not have a callback prop for this.

Test Plan: Passed in a callback which logged a message to the console as the `onClosed` prop and confirmed that the message appeared when the bottomsheet closed

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9098