[native] introduce user profile bottomsheet component
Summary: This diff introduces the `ProfileBottomSheet` component. When designing this component, I wanted to take a similar approach to how we handle popups/modals in the app today so I created a new screen in root navigator which we will navigate to when the user profile bottomsheet should appear. The "screen" we will navigate to is this component

Test Plan: Confirmed that I am able to navigate to and back from this new "screen"

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9101