[native] introduce user profile bottomsheet component
Summary:
This diff introduces the `ProfileBottomSheet` component. When designing this component/how bottomsheets should work in our app, I wanted to take a similar approach to how we currently handle popups/modals in the app today so I created a new screen in root navigator which we will navigate to when the user profile bottomsheet should appear. The "screen" we will navigate to is this `ProfileBottomSheet` component

Depends on D9098

Test Plan:
Confirmed that I am able to navigate to and back from this new "screen".

I created this test function in `RelationshipListItem`:

```
  onPressTestNav = userInfo => {
    this.props.navigate<'UserProfileBottomSheet'>({
      name: UserProfileBottomSheetRouteName,
      params: {
        userID: userInfo.id,
      },
    });
  };
```

and got the following:
{F746628}

Reviewers: atul, inka

Reviewed By: inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9101